refactor: GeminiApiException을 exception 패키지로, GeminiApiClient를 application 포트로 이동
